requirements.txt
azure-ai-textanalytics
beautifulsoup4 == 4.8.2
matplotlib == 3.3.4
nlpaug == 1.1.3
pandas == 1.1.5
scikit_learn == 0.24.2
boto3
google-cloud-language
